requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
pandas==1.5.3
openpyxl==3.0.10
Pillow>=10.0.0
pywin32>=305
cryptography>=41.0.0